fix: 修复 TypeScript 编译错误以支持 Vercel 部署
- 修复 BigInt 渲染错误 (DashboardV2.tsx:434)
- 统一 Page 类型定义，移除重复定义
- 修复 colors.text 属性访问错误
- 修复 useUserVeNFTs hooks 返回值属性名 (tokens -> nfts)
- 添加类型断言修复 Address 类型推断问题
- 清理所有未使用的变量和导入
- 移除未使用的组件 (InfoRow)

+++ frontend/src/hooks/useVeNFT.ts
@@ -212,7 +212,7 @@ export function useUserVeNFTs() {
   return {
     balance: balance as bigint | undefined,
     nfts,
-    isLoading: !!address && !!balance && balance > 0n && !lockedData,
+    isLoading: !!address && !!balance && (balance as bigint) > 0n && !lockedData,
   }
 }
